drop stream handler and log via telemetry
