+++ lib/plug/cowboy.ex
@@ -240,6 +240,8 @@ defmodule Plug.Cowboy do
         other -> :erlang.error({:badarg, [other]})
       end
 
+    :telemetry.attach(:plug_cowboy, [:plug_cowboy, :early_error], &handle_early_error/4, nil)
+
     apply(:cowboy, start, args(scheme, plug, opts, cowboy_options))
   end
 
@@ -303,7 +305,7 @@ defmodule Plug.Cowboy do
     [ref || build_ref(plug, scheme), transport_options, protocol_options]
   end
 
-  @default_stream_handlers [Plug.Cowboy.Telemetry, Plug.Cowboy.Stream]
+  @default_stream_handlers [Plug.Cowboy.Telemetry, :cowboy_stream_h]
 
   defp set_stream_handlers(opts) do
     compress = Keyword.get(opts, :compress)
@@ -336,6 +338,44 @@ defmodule Plug.Cowboy do
 
   defp fail(message) do
     raise ArgumentError, "could not start Cowboy2 adapter, " <> message
+  end
+
+  def handle_early_error([:plug_cowboy, :early_error], _, %{reason: reason, request: request}, _) do
+    case reason do
+      {:connection_error, :limit_reached, specific_reason} ->
+        Logger.error("""
+        Cowboy returned 431 because it was unable to parse the request headers.
+
+        This may happen because there are no headers, or there are too many headers
+        or the header name or value are too large (such as a large cookie).
+
+        More specific reason is:
+
+            #{inspect(specific_reason)}
+
+        You can customize those limits when configuring your http/https
+        server. The configuration option and default values are shown below:
+
+            protocol_options: [
+              max_header_name_length: 64,
+              max_header_value_length: 4096,
+              max_headers: 100
+            ]
+
+        Request info:
+
+            peer: #{format_peer(request.peer)}
+            method: #{request.method || "<unable to parse>"}
+            path: #{request.path || "<unable to parse>"}
+        """)
+
+      _ ->
+        :ok
+    end
+  end
+
+  defp format_peer({addr, port}) do
+    "#{:inet_parse.ntoa(addr)}:#{port}"
   end
 
   defp option_deprecation_warning(:acceptors),
